# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Second Hand API",
    "description": "This is an API document for our products\n",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "fauzanjr1@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "https://secondhand-be.herokuapp.com",
      "description": "Production"
    },
    {
      "url": "https://secondhand-be-test.herokuapp.com",
      "description": "Test"
    }
  ],
...
  "paths": {
...
    "/user-detail": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get All new get-user-detail",
        "description": "Returns a map of status codes to quantities",
        "operationId": "getUserDetail",
        "responses": {
          "200": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Add a new product",
        "description": "Returns a map of status codes to quantities",
        "operationId": "putUserDetail",
        "requestBody": {
          "description": "Update Products",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/userdetail_body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful to add product"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/profile": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get profile user",
        "description": "Returns user profile",
        "operationId": "getProfile",
        "responses": {
          "200": {
            "description": "successful operation"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "userdetail_body": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Anvathar Yiep Yiep"
          },
          "city": {
            "type": "string",
            "example": "Semarang"
          },
          "address": {
            "type": "string",
            "example": "Semarang Tengah, Jawa Tengah"
          },
          "phone": {
            "type": "string",
            "example": "089699964999"
          },
          "image_url": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}